services/meta-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication & security
python-jose[cryptography]==3.3.0
cryptography>=43.0.1
passlib[bcrypt]==1.7.4
bcrypt>=4.1.3
python-multipart==0.0.6

# Validation & parsing
pydantic==2.5.0
email-validator==2.1.0

# Email functionality
jinja2>=3.1.0
aiosmtplib>=3.0.0

# Development & testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
httpx==0.25.2
